* dev_master: (63 commits)
  编号：	127442 标题：	IOS】B站首页优惠活动未显示，后台设置是首页显示优惠活动 【普通】 优先级：	1 站点：	c126
  位置居中
  编号：	126946 标题：	【ios】是否能添加在线客服2 点击客服图标后 显示两个选项 【普通】
  编号：126113 标题：【原生IOS】对接 忘记提款密码验证
  编号：	128427 标题：	反馈原生iosAPP app首页 游戏大厅 聊天室底色都修改与图中一致，底部【首页】单独调整为红色的显示 【普通】
  编号：	128407 标题：	苹果】app打包域名换https://70333c.com/ 【普通】
  编号：	128399 标题：	c242 原生APP 苹果 首页颜色 换成红色同步安卓原生APP下面提供截图 【普通】
  去掉越南菜球球颜色
  编号：128233
  编号：128240 标题：IOS】利息宝转入转出背景页面与利息宝首页背景页面不统一（红圈标注）
  编号：128244 标题：IOS】利息宝转入转出页面的“总金额”改成“利息宝余额” ‘
  chat fix
  调整位置
  编号：	127267 标题：	【原生IOS】注单 【普通】
  1
  编号：	128100 标题：	ios把下图标注的内容修改成备注的字样 【普通】
  编号：	128098 标题：	iOS安卓把下图的金字塔的图片更换一下，换成下图文件 【普通】
  编号：	128101 标题：	IOS】香港六合彩麻烦调整为进入默认展示特码B 【普通】
  完成游戏列表搜索
  调整位置
  ...

# Conflicts:
#	UGBWApp.xcodeproj/project.pbxproj
#	ug/Classes/Funds/Controller/取款/UGWithdrawalViewController.storyboard
#	ug/Classes/Helper/FishUtility/define/AppDefine.m

+++ ug/Classes/Helper/FishUtility/Category/Utils/NS/NSString+Utils/NSString+Utils.m
@@ -395,7 +395,7 @@
     }
     return _NSString(@"%@****%@",
                      head > 0 ? [self substringToIndex:head] : @"",
-                     tail > 0 ? [self substringFromIndex:self.length-tail] : @"");
+                     tail > 0 ? [self substringFromIndex:(int)self.length-tail] : @"");
 }
